[flaky] Split test_failure_2 into 2 (#15688)
* split test failure

* format

* add bazel

* remove flaky tag

* format

* up

* rerun
