+++ python/ray/tests/test_failure_2.py
@@ -17,8 +17,8 @@ import ray._private.utils
 from ray.util.placement_group import placement_group
 import ray.ray_constants as ray_constants
 from ray.cluster_utils import Cluster
-from ray.test_utils import (wait_for_condition, SignalActor, init_error_pubsub,
-                            get_error_message, Semaphore)
+from ray.test_utils import (init_error_pubsub, get_error_message, Semaphore,
+                            wait_for_condition)
 
 
 def test_warning_for_infeasible_tasks(ray_start_regular, error_pubsub):
@@ -341,91 +341,6 @@ def test_raylet_crash_when_get(ray_start_regular):
     thread.join()
 
 
-def test_connect_with_disconnected_node(shutdown_only):
-    config = {
-        "num_heartbeats_timeout": 50,
-        "raylet_heartbeat_period_milliseconds": 10,
-    }
-    cluster = Cluster()
-    cluster.add_node(num_cpus=0, _system_config=config)
-    ray.init(address=cluster.address)
-    p = init_error_pubsub()
-    errors = get_error_message(p, 1, timeout=5)
-    print(errors)
-    assert len(errors) == 0
-    # This node is killed by SIGKILL, ray_monitor will mark it to dead.
-    dead_node = cluster.add_node(num_cpus=0)
-    cluster.remove_node(dead_node, allow_graceful=False)
-    errors = get_error_message(p, 1, ray_constants.REMOVED_NODE_ERROR)
-    assert len(errors) == 1
-    # This node is killed by SIGKILL, ray_monitor will mark it to dead.
-    dead_node = cluster.add_node(num_cpus=0)
-    cluster.remove_node(dead_node, allow_graceful=False)
-    errors = get_error_message(p, 1, ray_constants.REMOVED_NODE_ERROR)
-    assert len(errors) == 1
-    # This node is killed by SIGTERM, ray_monitor will not mark it again.
-    removing_node = cluster.add_node(num_cpus=0)
-    cluster.remove_node(removing_node, allow_graceful=True)
-    errors = get_error_message(p, 1, timeout=2)
-    assert len(errors) == 0
-    # There is no connection error to a dead node.
-    errors = get_error_message(p, 1, timeout=2)
-    assert len(errors) == 0
-    p.close()
-
-
-@pytest.mark.parametrize(
-    "ray_start_cluster_head", [{
-        "num_cpus": 5,
-        "object_store_memory": 10**8,
-    }],
-    indirect=True)
-def test_parallel_actor_fill_plasma_retry(ray_start_cluster_head):
-    @ray.remote
-    class LargeMemoryActor:
-        def some_expensive_task(self):
-            return np.zeros(10**8 // 2, dtype=np.uint8)
-
-    actors = [LargeMemoryActor.remote() for _ in range(5)]
-    for _ in range(5):
-        pending = [a.some_expensive_task.remote() for a in actors]
-        while pending:
-            [done], pending = ray.wait(pending, num_returns=1)
-
-
-def test_fill_object_store_exception(shutdown_only):
-    ray.init(
-        num_cpus=2,
-        object_store_memory=10**8,
-        _system_config={"automatic_object_spilling_enabled": False})
-
-    @ray.remote
-    def expensive_task():
-        return np.zeros((10**8) // 10, dtype=np.uint8)
-
-    with pytest.raises(ray.exceptions.RayTaskError) as e:
-        ray.get([expensive_task.remote() for _ in range(20)])
-        with pytest.raises(ray.exceptions.ObjectStoreFullError):
-            raise e.as_instanceof_cause()
-
-    @ray.remote
-    class LargeMemoryActor:
-        def some_expensive_task(self):
-            return np.zeros(10**8 + 2, dtype=np.uint8)
-
-        def test(self):
-            return 1
-
-    actor = LargeMemoryActor.remote()
-    with pytest.raises(ray.exceptions.RayTaskError):
-        ray.get(actor.some_expensive_task.remote())
-    # Make sure actor does not die
-    ray.get(actor.test.remote())
-
-    with pytest.raises(ray.exceptions.ObjectStoreFullError):
-        ray.put(np.zeros(10**8 + 2, dtype=np.uint8))
-
-
 @pytest.mark.parametrize(
     "ray_start_cluster", [{
         "num_nodes": 1,
@@ -609,76 +524,6 @@ def test_gcs_server_failiure_report(ray_start_regular, log_pubsub):
             continue
         data = json.loads(ray._private.utils.decode(msg["data"]))
         assert data["pid"] == "gcs_server"
-
-
-@pytest.mark.parametrize(
-    "ray_start_regular", [{
-        "_system_config": {
-            "task_retry_delay_ms": 500
-        }
-    }],
-    indirect=True)
-def test_async_actor_task_retries(ray_start_regular):
-    # https://github.com/ray-project/ray/issues/11683
-
-    signal = SignalActor.remote()
-
-    @ray.remote
-    class DyingActor:
-        def __init__(self):
-            print("DyingActor init called")
-            self.should_exit = False
-
-        def set_should_exit(self):
-            print("DyingActor.set_should_exit called")
-            self.should_exit = True
-
-        async def get(self, x, wait=False):
-            print(f"DyingActor.get called with x={x}, wait={wait}")
-            if self.should_exit:
-                os._exit(0)
-            if wait:
-                await signal.wait.remote()
-            return x
-
-    # Normal in order actor task retries should work
-    dying = DyingActor.options(
-        max_restarts=-1,
-        max_task_retries=-1,
-    ).remote()
-
-    assert ray.get(dying.get.remote(1)) == 1
-    ray.get(dying.set_should_exit.remote())
-    assert ray.get(dying.get.remote(42)) == 42
-
-    # Now let's try out of order retries:
-    # Task seqno 0 will return
-    # Task seqno 1 will be pending and retried later
-    # Task seqno 2 will return
-    # Task seqno 3 will crash the actor and retried later
-    dying = DyingActor.options(
-        max_restarts=-1,
-        max_task_retries=-1,
-    ).remote()
-
-    # seqno 0
-    ref_0 = dying.get.remote(0)
-    assert ray.get(ref_0) == 0
-    # seqno 1
-    ref_1 = dying.get.remote(1, wait=True)
-    # seqno 2
-    ref_2 = dying.set_should_exit.remote()
-    assert ray.get(ref_2) is None
-    # seqno 3, this will crash the actor because previous task set should exit
-    # to true.
-    ref_3 = dying.get.remote(3)
-
-    # At this point the actor should be restarted. The two pending tasks
-    # [ref_1, ref_3] should be retried, but not the completed tasks [ref_0,
-    # ref_2]. Critically, if ref_2 was retried, ref_3 can never return.
-    ray.get(signal.send.remote())
-    assert ray.get(ref_1) == 1
-    assert ray.get(ref_3) == 3
 
 
 def test_raylet_node_manager_server_failure(ray_start_cluster_head,
